refactor: make some files pass in strict mode
As a proof of concept, I want to make sure that npm run test is able to
build the code in strict mode. This already triggers some errors on some
files that are required during the test run. Not every file fails
because the coverage of this project is not high.

I've fixed those issues.

Also, I've fixed any issue in src/cmd/*.ts as reported by my text
editor.

+++ src/cmd/makibotctl.ts
@@ -110,18 +110,21 @@ makibotctl.command<{ app: string; local: string }>(
   "Remove deployed commands for a local guild",
   () => ({}),
   async (argv) => {
+    if (!process.env.BOT_TOKEN) {
+      console.error("Please, provide the bot token in the BOT_TOKEN env variable");
+      process.exit(1);
+    }
+
     const restClient = new REST({ version: "9" });
     restClient.setToken(process.env.BOT_TOKEN);
 
     /* Fetch the commands. */
-    restClient
-      .get(Routes.applicationGuildCommands(argv.app, argv.local))
-      .then(async (commands: RESTGetAPIApplicationGuildCommandsResult) => {
-        for (const command of commands) {
-          console.log(`Deleting local command ${command.name}`);
-          await restClient.delete(Routes.applicationGuildCommand(argv.app, argv.local, command.id));
-        }
-      });
+    const url = Routes.applicationGuildCommands(argv.app, argv.local);
+    const response = (await restClient.get(url)) as RESTGetAPIApplicationGuildCommandsResult;
+    for (const command of response) {
+      console.log(`Deleting local command ${command.name}`);
+      await restClient.delete(Routes.applicationGuildCommand(argv.app, argv.local, command.id));
+    }
   },
 );
